* Fixes tests  - Fixes a gazillion tests  - Revert to Boot 2.0.0.M3. Snapshots breaking actuator  - Checkstyle fixes  - Disable JsonUnmarshalling as a catch all converter
