+++ pom.xml
@@ -9,7 +9,7 @@
     <parent>
         <groupId>org.springframework.cloud</groupId>
         <artifactId>spring-cloud-build</artifactId>
-        <version>2.0.0.BUILD-SNAPSHOT</version>
+        <version>2.0.0.M2</version>
         <relativePath/>
     </parent>
     <scm>
