+++ spring-cloud-stream-integration-tests/src/test/java/org/springframework/cloud/stream/config/TextPlainConversionTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.springframework.cloud.stream.config.broken;
+package org.springframework.cloud.stream.config;
 
 import java.util.concurrent.TimeUnit;
 
@@ -25,26 +25,26 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.cloud.stream.annotation.EnableBinding;
-import org.springframework.cloud.stream.annotation.StreamListener;
 import org.springframework.cloud.stream.binder.BinderFactory;
 import org.springframework.cloud.stream.messaging.Processor;
 import org.springframework.cloud.stream.test.binder.TestSupportBinder;
+import org.springframework.context.annotation.PropertySource;
+import org.springframework.integration.annotation.ServiceActivator;
 import org.springframework.integration.support.MessageBuilder;
 import org.springframework.messaging.Message;
 import org.springframework.messaging.MessageChannel;
-import org.springframework.messaging.MessageHeaders;
-import org.springframework.messaging.handler.annotation.SendTo;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 /**
  * @author Marius Bogoevici
+ *
  * @since 1.2
  */
 @RunWith(SpringJUnit4ClassRunner.class)
-@SpringBootTest(classes = TextPlainToJsonConversionTest.FooProcessor.class, webEnvironment = SpringBootTest.WebEnvironment.NONE)
-public class TextPlainToJsonConversionTest {
+@SpringBootTest(classes = TextPlainConversionTest.FooProcessor.class, webEnvironment = SpringBootTest.WebEnvironment.NONE)
+public class TextPlainConversionTest {
 
 	@Autowired
 	private Processor testProcessor;
@@ -53,34 +53,43 @@ public class TextPlainToJsonConversionTest {
 	private BinderFactory binderFactory;
 
 	@Test
-	public void testNoContentTypeToJsonConversionOnInput() throws Exception {
-		testProcessor.input().send(MessageBuilder.withPayload("{\"name\":\"Bar\"}").build());
-		Message<?> received = ((TestSupportBinder) binderFactory.getBinder(null, MessageChannel.class))
+	public void testTextPlainConversionOnOutput() throws Exception {
+		testProcessor.input().send(MessageBuilder.withPayload("Bar").build());
+		@SuppressWarnings("unchecked")
+		Message<byte[]> received = (Message<byte[]>) ((TestSupportBinder) binderFactory.getBinder(null, MessageChannel.class))
 				.messageCollector().forChannel(testProcessor.output()).poll(1, TimeUnit.SECONDS);
 		assertThat(received).isNotNull();
-		assertThat(((Foo) received.getPayload()).getName()).isEqualTo("transformed-Bar");
+		assertThat(new String(received.getPayload())).isEqualTo("Foo{name='Bar'}");
 	}
 
 	@Test
-	public void testTextPlainToJsonConversionOnInput() throws Exception {
-		testProcessor.input().send(MessageBuilder.withPayload("{\"name\":\"Bar\"}")
-				.setHeader(MessageHeaders.CONTENT_TYPE, "text/plain").build());
-		Message<?> received = ((TestSupportBinder) binderFactory.getBinder(null, MessageChannel.class))
+	public void testByteArrayConversionOnOutput() throws Exception {
+		testProcessor.output().send(MessageBuilder.withPayload("Bar".getBytes()).build());
+		@SuppressWarnings("unchecked")
+		Message<byte[]> received = (Message<byte[]>) ((TestSupportBinder) binderFactory.getBinder(null, MessageChannel.class))
 				.messageCollector().forChannel(testProcessor.output()).poll(1, TimeUnit.SECONDS);
 		assertThat(received).isNotNull();
-		assertThat(((Foo) received.getPayload()).getName()).isEqualTo("transformed-Bar");
+		assertThat(new String(received.getPayload())).isEqualTo("Bar");
+	}
+
+	@Test
+	public void testTextPlainConversionOnInputAndOutput() throws Exception {
+		testProcessor.input().send(MessageBuilder.withPayload(new Foo("Bar")).build());
+		@SuppressWarnings("unchecked")
+		Message<byte[]> received = (Message<byte[]>) ((TestSupportBinder) binderFactory.getBinder(null, MessageChannel.class))
+				.messageCollector().forChannel(testProcessor.output()).poll(1, TimeUnit.SECONDS);
+		assertThat(received).isNotNull();
+		assertThat(new String(received.getPayload())).isEqualTo("Foo{name='Foo{name='Bar'}'}");
 	}
 
 	@EnableBinding(Processor.class)
 	@EnableAutoConfiguration
+	@PropertySource("classpath:/org/springframework/cloud/stream/config/textplain/text-plain.properties")
 	public static class FooProcessor {
 
-		@StreamListener("input")
-		@SendTo("output")
-		public Foo consume(Foo foo) {
-			Foo returnFoo = new Foo();
-			returnFoo.setName("transformed-" + foo.getName());
-			return returnFoo;
+		@ServiceActivator(inputChannel = "input", outputChannel = "output")
+		public Foo consume(String foo) {
+			return new Foo(foo);
 		}
 
 	}
@@ -89,7 +98,8 @@ public class TextPlainToJsonConversionTest {
 
 		private String name;
 
-		public Foo() {
+		public Foo(String name) {
+			this.name = name;
 		}
 
 		public String getName() {
