+++ spring-cloud-stream-integration-tests/src/test/java/org/springframework/cloud/stream/config/broken/StreamListenerHandlerMethodTests.java
@@ -65,6 +65,7 @@ import static org.springframework.cloud.stream.binding.StreamListenerErrorMessag
  * @author Marius Bogoevici
  * @author Ilayaperumal Gopinathan
  * @author Gary Russell
+ * @author Vinicius Carvalho
  */
 public class StreamListenerHandlerMethodTests {
 
@@ -82,30 +83,35 @@ public class StreamListenerHandlerMethodTests {
 	@Test
 	public void testMethodWithObjectAsMethodArgument() throws Exception {
 		ConfigurableApplicationContext context = SpringApplication.run(TestMethodWithObjectAsMethodArgument.class,
-				"--server.port=0");
+				"--server.port=0","--spring.cloud.stream.bindings.input.contentType=text/plain","--spring.cloud.stream.bindings.output.contentType=text/plain");
 		Processor processor = context.getBean(Processor.class);
 		final String testMessage = "testing";
 		processor.input().send(MessageBuilder.withPayload(testMessage).build());
 		MessageCollector messageCollector = context.getBean(MessageCollector.class);
-		Message<?> result = messageCollector.forChannel(processor.output()).poll(1000, TimeUnit.MILLISECONDS);
+		Message<byte[]> result = (Message<byte[]>) messageCollector.forChannel(processor.output()).poll(1000, TimeUnit.MILLISECONDS);
 		assertThat(result).isNotNull();
-		assertThat(result.getPayload()).isEqualTo(testMessage.toUpperCase());
+		assertThat(new String(result.getPayload())).isEqualTo(testMessage.toUpperCase());
 		context.close();
 	}
 
 	@Test
+	/**
+	 * @since 2.0 : This test is an example of the new behavior of 2.0 when it comes to contentType handling.
+	 * The default contentType being JSON in order to be able to check a message without quotes the user needs to set the input/output contentType accordingly
+	 * Also, received messages are always of Message<byte[]> now.
+	 */
 	public void testMethodHeadersPropagatged() throws Exception {
 		ConfigurableApplicationContext context = SpringApplication.run(TestMethodHeadersPropagated.class,
-				"--server.port=0");
+				"--server.port=0","--spring.cloud.stream.bindings.input.contentType=text/plain","--spring.cloud.stream.bindings.output.contentType=text/plain");
 		Processor processor = context.getBean(Processor.class);
 		final String testMessage = "testing";
 		processor.input().send(MessageBuilder.withPayload(testMessage)
 				.setHeader("foo", "bar")
 				.build());
 		MessageCollector messageCollector = context.getBean(MessageCollector.class);
-		Message<?> result = messageCollector.forChannel(processor.output()).poll(1000, TimeUnit.MILLISECONDS);
+		Message<byte[]> result = (Message<byte[]>) messageCollector.forChannel(processor.output()).poll(1000, TimeUnit.MILLISECONDS);
 		assertThat(result).isNotNull();
-		assertThat(result.getPayload()).isEqualTo(testMessage.toUpperCase());
+		assertThat(new String(result.getPayload())).isEqualTo(testMessage.toUpperCase());
 		assertThat(result.getHeaders().get("foo")).isEqualTo("bar");
 		context.close();
 	}
@@ -113,34 +119,35 @@ public class StreamListenerHandlerMethodTests {
 	@Test
 	public void testMethodHeadersNotPropagatged() throws Exception {
 		ConfigurableApplicationContext context = SpringApplication.run(TestMethodHeadersNotPropagated.class,
-				"--server.port=0");
+				"--server.port=0","--spring.cloud.stream.bindings.input.contentType=text/plain","--spring.cloud.stream.bindings.output.contentType=text/plain");
 		Processor processor = context.getBean(Processor.class);
 		final String testMessage = "testing";
 		processor.input().send(MessageBuilder.withPayload(testMessage)
 				.setHeader("foo", "bar")
 				.build());
 		MessageCollector messageCollector = context.getBean(MessageCollector.class);
-		Message<?> result = messageCollector.forChannel(processor.output()).poll(1000, TimeUnit.MILLISECONDS);
+		Message<byte[]> result = (Message<byte[]>) messageCollector.forChannel(processor.output()).poll(1000, TimeUnit.MILLISECONDS);
 		assertThat(result).isNotNull();
-		assertThat(result.getPayload()).isEqualTo(testMessage.toUpperCase());
+		assertThat(new String(result.getPayload())).isEqualTo(testMessage.toUpperCase());
 		assertThat(result.getHeaders().get("foo")).isNull();
 		context.close();
 	}
 
 	@Test
+	//TODO: Handle dynamic destinations and contentType
 	public void testStreamListenerMethodWithTargetBeanFromOutside() throws Exception {
 		ConfigurableApplicationContext context = SpringApplication
-				.run(TestStreamListenerMethodWithTargetBeanFromOutside.class, "--server.port=0");
+				.run(TestStreamListenerMethodWithTargetBeanFromOutside.class, "--server.port=0","--spring.cloud.stream.bindings.input.contentType=text/plain","--spring.cloud.stream.bindings.output.contentType=text/plain");
 		Sink sink = context.getBean(Sink.class);
 		final String testMessageToSend = "testing";
 		sink.input().send(MessageBuilder.withPayload(testMessageToSend).build());
 		DirectChannel directChannel = (DirectChannel) context.getBean(testMessageToSend.toUpperCase(),
 				MessageChannel.class);
 		MessageCollector messageCollector = context.getBean(MessageCollector.class);
-		Message<?> result = messageCollector.forChannel(directChannel).poll(1000, TimeUnit.MILLISECONDS);
+		Message<byte[]> result = (Message<byte[]>) messageCollector.forChannel(directChannel).poll(1000, TimeUnit.MILLISECONDS);
 		sink.input().send(MessageBuilder.withPayload(testMessageToSend).build());
 		assertThat(result).isNotNull();
-		assertThat(result.getPayload()).isEqualTo(testMessageToSend.toUpperCase());
+		assertThat(new String(result.getPayload())).isEqualTo(testMessageToSend.toUpperCase());
 		context.close();
 	}
