+++ spring-cloud-stream-integration-tests/src/test/resources/org/springframework/cloud/stream/config/channel/partitioned-configurers.properties
@@ -1,3 +1,4 @@
 spring.cloud.stream.bindings.output.destination=partOut
 spring.cloud.stream.bindings.output.producer.partitionKeyExpression=payload
 spring.cloud.stream.bindings.output.producer.partitionCount=3
+
