+++ spring-cloud-stream/src/main/java/org/springframework/cloud/stream/annotation/StreamConverter.java
@@ -13,6 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+
 package org.springframework.cloud.stream.annotation;
 
 import java.lang.annotation.Documented;
