+++ spring-cloud-stream/src/test/java/org/springframework/cloud/stream/binder/ExtendedPropertiesBinderAwareChannelResolverTests.java
@@ -135,7 +135,7 @@ public class ExtendedPropertiesBinderAwareChannelResolverTests extends BinderAwa
 			fail("interrupted while awaiting latch");
 		}
 		assertThat(received).hasSize(1);
-		assertThat(received.get(0).getPayload()).isEqualTo("hello");
+		assertThat(new String((byte[])received.get(0).getPayload())).isEqualTo("hello");
 		context.close();
 		assertThat(producerBindings).hasSize(1);
 		assertThat(binding.isBound()).isFalse().describedAs("Must not be bound");
